fix: rename validation/types.py to models.py to avoid stdlib shadow
scripts/validation/types.py shadows Python's stdlib `types` module.
When system python3 imports from scripts/validation/, the stdlib
`dataclasses` -> `copy` -> `weakref` -> `types` chain resolves to
the local file instead, causing a circular ImportError.

Rename to models.py and update all 6 import sites.

Fixes #1387
