+++ scripts/validation/__init__.py
@@ -3,6 +3,6 @@
 Migrated from PowerShell Validate-*.ps1 scripts per ADR-042.
 """
 
-from scripts.validation.types import ValidationResult
+from scripts.validation.models import ValidationResult
 
 __all__ = ["ValidationResult"]
